refactor: create verify submodule
